+++ test.sh
@@ -37,7 +37,11 @@ function build_app {
     # test install from binary with fallback
     # run directly against node-pre-gyp
     node-pre-gyp clean
-    node-pre-gyp install --fallback-to-build $2
+    if [[ $1  == "app2" ]]; then
+        node-pre-gyp install --fallback-to-build --custom_include_path=`pwd`/include
+    else
+        node-pre-gyp install --fallback-to-build
+    fi
     npm test
 
     if [[ $TRAVIS_PULL_REQUEST == true ]] ; then
@@ -48,13 +52,15 @@ function build_app {
     else
         MARK "D" $1
         # it works, so now publish
-        node-pre-gyp package publish $2
+        node-pre-gyp package publish
         node-pre-gyp testpackage --overwrite
+        node-pre-gyp unpublish
+        node-pre-gyp publish
 
         MARK "E" $1
         # now test installing via remote binary without fallback
         node-pre-gyp clean
-        npm install --fallback-to-build=false $2
+        npm install --fallback-to-build=false
         npm test
     fi
 
@@ -63,7 +69,11 @@ function build_app {
     for i in $(find . -name '*.node') ; do
         echo 'bogus' > $i;
     done
-    npm install $2
+    if [[ $1  == "app2" ]]; then
+        npm install --custom_include_path=`pwd`/include
+    else
+        npm install
+    fi
     npm test
 
     # cleanup
@@ -75,11 +85,11 @@ function build_app {
 
 setup
 # simpliest, least config node c++ addon possible
-build_app "app1" ""
+build_app "app1"
 # app with more custom organization and needing a variable passed for custom include path
-build_app "app2" "--custom_include_path=`pwd`/app2/include"
+build_app "app2"
 # app that depends on an external static library
-build_app "app3" ""
+build_app "app3"
 # app that depends on an external shared library
-build_app "app4" ""
+build_app "app4"
 teardown